fixing delete on file and not directory

+++ src/main/java/com/ibm/stocator/fs/ObjectStoreFileSystem.java
@@ -335,7 +335,8 @@ public class ObjectStoreFileSystem extends ExtendedFileSystem {
         }
       }
       LOG.debug("Delete {} the root", pathToObj);
-      if (f.getParent() != null && !f.getParent().toString().equals(hostNameScheme)) {
+      if (!fsT.isDirectory() && f.getParent() != null
+          && !f.getParent().toString().equals(hostNameScheme)) {
         storageClient.delete(hostNameScheme, f, false);
       }
     }
